added pwm class and header
