+++ led/pwm.h
@@ -1,0 +1,7 @@
+#define PWM_H
+
+#include <stdint.h>
+#include <stdio.h>
+#include "nrf_drv_pwm.h"
+
+void power_leds(uint32_t num_leds);
